# contracts/openapi/admin.yaml
openapi: 3.1.0
info:
  title: RingSays Admin API
  version: 0.2.0
  description: |
    API behind the enterprise portal and the RingSays back office.

    Authentication: OpenID Connect. The identity provider proves who the person is (`iss`, `sub`, verified
    email). RingSays decides what they may do: tenant memberships and roles are stored by RingSays, so a
    role change or a disabled user takes effect on the next request, not at token expiry.

    Tenant roles: TENANT_ADMIN, INTEGRATION_ADMIN, SUPERVISOR, AGENT, COMPLIANCE.
    RingSays staff roles: RS_REVIEWER (decides reviews), RS_ADMIN (creates and suspends tenants).

    Containment stays available while an organisation is suspended: revoking credentials, disabling
    webhook endpoints, stopping a calling number and disabling people.

    Back office paths (`/backoffice/...`, `/review/...`) are served only by the internal back office
    deployment, which holds the back office database role. The tenant facing deployment answers 404.

    Every write is recorded in the tenant's hash chained audit log with actor and, where required, reason.
    Phone numbers are never returned in full by this API.
servers:
  - url: https://api.ringsays.example/admin/v1
security:
  - oidc: []
...
paths:
  /me:
    get:
      tags: [Session]
      operationId: getMe
      summary: Signed in person, tenant memberships and staff roles
      description: First sign in with a verified email activates pending invitations for that email.
      responses:
        "200":
          description: Session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "./common.yaml#/components/responses/Unauthorized" }

  # Organisation
  /tenants/{tenant_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      tags: [Organisation]
      operationId: getTenant
      summary: Get tenant profile
      responses:
        "200":
          description: Tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
    patch:
      tags: [Organisation]
      operationId: updateTenant
      summary: Update profile (only before verification is submitted)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantUpdate" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "400": { $ref: "./common.yaml#/components/responses/BadRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
  /tenants/{tenant_id}/departments:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      tags: [Organisation]
      operationId: listDepartments
      summary: List departments
      responses:
        "200":
          description: Departments.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Department" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
    post:
      tags: [Organisation]
      operationId: createDepartment
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DepartmentInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "400": { $ref: "./common.yaml#/components/responses/BadRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
  /tenants/{tenant_id}/departments/{department_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: department_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Organisation]
      operationId: updateDepartment
      summary: Rename department
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DepartmentInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
  /tenants/{tenant_id}/agents:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      tags: [Organisation]
      operationId: listAgents
      summary: List agents
      parameters:
        - $ref: "./common.yaml#/components/parameters/Cursor"
        - $ref: "./common.yaml#/components/parameters/Limit"
      responses:
        "200":
          description: Agents.
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
                  next_cursor: { type: [string, "null"] }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
    post:
      tags: [Organisation]
      operationId: createAgent
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "./common.yaml#/components/responses/BadRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }
  /tenants/{tenant_id}/agents/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: agent_id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Organisation]
      operationId: updateAgent
      summary: Update or deactivate agent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentUpdate" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }
  /tenants/{tenant_id}/calling-numbers:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      tags: [Organisation]
      operationId: listCallingNumbers
      summary: List calling numbers
      responses:
        "200":
          description: Calling numbers.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CallingNumber" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
    post:
      tags: [Organisation]
      operationId: addCallingNumber
      summary: Add calling number for verification
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CallingNumberInput" }
      responses:
        "201":
          description: Added, status PENDING_VERIFICATION.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CallingNumber" }
        "400": { $ref: "./common.yaml#/components/responses/BadRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }
  /tenants/{tenant_id}/calling-numbers/{number_id}/revoke:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/NumberId"
    post:
      tags: [Organisation]
      operationId: revokeCallingNumber
      summary: Stop using a calling number
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Reason" }
      responses:
        "200":
          description: Revoked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CallingNumber" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }

  # Verification
  /tenants/{tenant_id}/verification:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      tags: [Verification]
      operationId: getVerification
      summary: Verification evidence and latest request
      responses:
        "200":
          description: Verification state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Verification" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
  /tenants/{tenant_id}/verification/documents:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    post:
      tags: [Verification]
      operationId: uploadVerificationDocument
      summary: Upload evidence (PDF, PNG or JPEG, at most 5 MB)
      description: Type is read from the file content. Requests above the size limit are refused before authentication.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [kind, file]
              properties:
                kind: { $ref: "#/components/schemas/DocumentKind" }
                reference: { type: string, maxLength: 80 }
                file: { type: string, contentMediaType: application/octet-stream }
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationDocument" }
        "400": { $ref: "./common.yaml#/components/responses/BadRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }
  /tenants/{tenant_id}/verification/documents/{document_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/DocumentId"
    delete:
      tags: [Verification]
      operationId: deleteVerificationDocument
      summary: Remove evidence (not while a request is under review)
      responses:
        "204": { description: Removed. }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
  /tenants/{tenant_id}/verification/submit:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    post:
      tags: [Verification]
      operationId: submitVerification
      summary: Submit evidence for RingSays review
      responses:
        "201":
          description: Submitted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }

  # Users
...
  /tenants/{tenant_id}/users/{user_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: user_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Users]
      operationId: updatePortalUser
      summary: Change roles, agent link or status
      description: |
        Nobody may change their own roles or disable themselves; another administrator must.
        While the organisation is suspended only `{"status": "DISABLED"}` is accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortalUserUpdate" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortalUser" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }

  # Catalogue
...
components:
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.ringsays.example/.well-known/openid-configuration
  parameters:
    TenantId:
      name: tenant_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    NumberId:
      name: number_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    TenantUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        legal_name: { $ref: "./common.yaml#/components/schemas/LocalisedText" }
        commercial_registration: { type: string, pattern: '^[0-9]{10}$' }
        domain:
          type: string
          maxLength: 253
          pattern: '^([A-Za-z0-9]([A-Za-z0-9-]{0,61}[A-Za-z0-9])?\.)+[A-Za-z]{2,63}$'
...